memcached-image: Drop xslt/xsl and use a jinja2 template with j2gen
xstl is complicated especially when trying to have nested
configuration setups (like having a common config for all containers,
a common config for image specific containers and a config for distro
specific containers) which need to be merged together before creating
the final .kiwi file.
Switch to j2gen[1] which is a generic Jinja2 (a python based template
engine) renderer. That way, the setup now have multiple files:

- common/global.yml - contains variables that are used for *all* container
  images
- memcached-image/common.yml - contains variables that are used for
  all memcached images
- memcached-image/openSUSE_Tumbleweed.yml - contains variables that
  are used only for the memcached openSUSE Tumbleweed based container
- memcached-image/memcached-image.kiwi.j2 - the kiwi file that is a
  Jinja2 template and can be processed (together with the other files)
  with j2gen

Generating the memcached-image.kiwi file can be done with:

j2gen generate -o memcached-image.kiwi \
  memcached-image/memcached-image.kiwi.j2 \
  common/global.yml memcached-image/common.yml \
  memcached-image/openSUSE_Tumbleweed.yml

Note: The ordering of the give files is important for j2gen.

Also use a root/ dir instead of an archive to inject files into the
memcached container. kiwi is copying the root/ directory
automatically.

Note: docker-entrypoint.sh is a hardlink to the submodule
file. Softlinks will not work here because we don't copy the target
into the container.x

[1] https://pypi.org/project/j2gen/